# packages/agile-security/docs/api/swagger/entity-api-swagger.yml
swagger: '2.0'
info:
  title: Agile IDM (Entity) HTTP API
  description: The Agile HTTP API to interact with AGILE IDM through http(s)
  version: 1.0.0
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# the domain of the service
host: localhost:3000
...
schemes:
  - http
# will be prefixed to all paths
basePath: /api/v1
consumes:
  - application/json
produces:
  - application/json
...
securityDefinitions:
  Oauth2AuthCode:
    type: oauth2
    flow: accessCode
    authorizationUrl: /oauth2/dialog/authorize
    tokenUrl: /oauth/token
  Oauth2Implicit:
    type: oauth2
    flow: implicit
    authorizationUrl: /oauth2/dialog/authorize
  Oauth2ClientCred:
    type: oauth2
    flow: application
    tokenUrl: /oauth2/token
paths:
  '/user/':

    get:
      security:
      -  Oauth2AuthCode: []
      - Oauth2Implicit: []
      - Oauth2ClientCred: []
      parameters:
      -
        name: auth_type
        in: query
        description: 'Authentication type, i.e. local, github, google'
        required: true
        type: string
      -
        name: user_name
        in: query
        description: 'username'
        required: true
        type: string
      operationId: ReadUser
      tags:
        - User
      description: 'Return a given user  by authentication type and username.'
      responses:
        '200':
          description: 'User read done'
          schema:
            $ref: '#/definitions/User'
          examples:
            application/json:
              id: bob!@!local
              type: /user
              owner: bob!@!local
              user_name: bob
              auth_type: local
        '401':
          description: 'Not authenticated'
        '403':
          description: Forbidden
        '404':
          description: 'No user found'
        '500':
          description: 'Unexpected error'

    post:
      security:
      - Oauth2AuthCode: []
      - Oauth2Implicit: []
      - Oauth2ClientCred: []
      operationId: StoreUser
      tags:
        - User
      description: 'Store a given user'
      parameters:
        -
          name: entity
          in: body
          description: 'the user to be stored'
          required: true
          schema:
            $ref: '#/definitions/User'

      responses:
        '200':
          description: 'Usertored successfully'
          schema:
             $ref: '#/definitions/User'
          examples:
            application/json:
              id: bob!@!local
              type: /user
              owner: bob!@!local
              user_name: bob
              auth_type: local
        '401':
          description: 'Not authenticated'
        '403':
          description: Forbidden
        '409':
          description: 'Conflict. This can happen when entities already exist for example'
        '422':
          description: 'This is returned when users are attempted to be created through this API, for this there is a user api.'
        '500':
          description: 'Unexpected error'

    delete:
      security:
      -  Oauth2AuthCode: []
      - Oauth2Implicit: []
      - Oauth2ClientCred: []
      parameters:
      -
        name: auth_type
        in: query
        description: 'Authentication type, i.e. local, github, google'
        required: true
        type: string
      -
        name: user_name
        in: query
        description: 'username'
        required: true
        type: string
      operationId: DeleteUser
      tags:
        - User
      description: 'Delete given user  by authentication type and username.'
      responses:
        '200':
          description: 'User delete done'
        '401':
          description: 'Not authenticated'
        '403':
          description: Forbidden
        '404':
          description: 'No user found'
        '500':
          description: 'Unexpected error'
  '/entity/{entityType}/{entityId}':
    parameters:
      -
        name: entityType
        in: path
        description: 'Entity type'
        required: true
        type: string
      -
        name: entityId
        in: path
        description: 'Entity id. This id is only unique for a certain entity type'
        required: true
        type: string
    get:
      security:
      -  Oauth2AuthCode: []
      - Oauth2Implicit: []
      - Oauth2ClientCred: []

      operationId: ReadEntity
      tags:
        - Entity
      description: 'Return a given entity by type and id.'
      responses:
        '200':
          description: 'Entity read done'
          schema:
            $ref: '#/definitions/Entity'
          examples:
            application/json:
              id: bleA0E6F8B62304
              type: /user
              owner: bob!@!local
              user_name: bob
              auth_type: local
        '401':
          description: 'Not authenticated'
        '403':
          description: Forbidden
        '404':
          description: 'No entity found'
        '500':
          description: 'Unexpected error'
    post:
      security:
      - Oauth2AuthCode: []
      - Oauth2Implicit: []
      - Oauth2ClientCred: []
      operationId: StoreEntity
      tags:
        - Entity
      description: 'Store a given entity under an id and type.'
      parameters:
        -
          name: entity
          in: body
          description: 'the entity to be stored'
          required: true
          schema:
            type: object

      responses:
        '200':
          description: 'Entity stored successfully'
          schema:
            type: object
          examples:
            application/json:
              id: bleA0E6F8B62304
              type: /user
              owner: bob!@!local
              user_name: bob
              auth_type: local
        '401':
          description: 'Not authenticated'
        '403':
          description: Forbidden
        '409':
          description: 'Conflict. This can happen when entities already exist for example'
        '422':
          description: 'This is returned when users are attempted to be created through this API, for this there is a user api.'
        '500':
          description: 'Unexpected error'
    delete:
      security:
        -  Oauth2AuthCode: []
        - Oauth2Implicit: []
        - Oauth2ClientCred: []
      operationId: DeleteEntity
      tags:
        - Entity
      description: 'Delete a given entity under an id and type.'
      responses:
        '200':
          description: 'Entity deleted successfully'
        '401':
          description: 'Not authenticated'
        '403':
          description: Forbidden
        '404':
          description: 'Entity not found'
        '500':
          description: 'Unexpected error'
...
definitions:
...
  Entity:
    description: 'An AGILE IDM Entity. This model shows only fields placed in the entity by AGILE. In addition to these fields, each object will have the attributes provided when the entity was created.'
    type: object
    properties:
      id:
        type: string
      type:
        type: string
      owner:
        type: string
...
  User:
    description: 'An AGILE User is also an Entity. This model shows the mandatory fields in a user, but this can be extended via the entity definition in the configuration file'
    type: object
    properties:
      id:
        type: string
      auth_type:
        type: string
      user_name:
        type: string
      owner:
        type: string